feat: add normalize_semver to semver_utils, drop duplicate module
